Add a quickly accessible journal
This commit creates a quick access journal with a focused edit mode. The
journal can always be pulled up with '<leader>j' and opens with Goyo and
Limelight for a clean interface. You can just as easily return to the
buffer your were working on with the same keystroke.

A number of colorscheme changes had to be made to support the different
formatting of the journal.