Return to last tab after exiting journal
This commit fixes an issue where exiting the journal would return you to
the tab to the right of your last tab, breaking the flow of quickly
editing the journal. This commit adds functionality to store your tab
number and use it to always return to the correct tab on closing the
journal.