Add a fast close back to shell from journal
This commit adds a flag to Vim when the journal is opened with the 'j'
command from the shell. If this flag is found when ToggleJournal() is
called, then Vim is closed entirely instead of ending on an empty
buffer.  This allows for a more natural workflow where, if you quickly
open the journal from a shell, toggling it closed returns you back to
the shell just as quickly.