Install critical Python modules during bootstrap
The Python module `neovim` is requred to use the `ddev theme` utility
(it is used to automatically sync the theme in all Neovim instances).
This commit adds a `pip install` of this module as part of the bootstrap
script. Additionally, the commit switches the Python command from
`python3` to `/usr/bin/python3` so that the `neovim` module will be
available even if we are within a Python virtual environment that might
not have the module installed.